Intermediate fixes final_query_3-23.ipynb
AttributeError encountered adding the WDS columns to the query_results_table. The code previously used by Daphne was:

 ```
# add the wds identifier column and source id columns (doesn't work until I add one row to the table)
            query_results_table.add_row()
            query_results_table.add_column('                              ', name = 'wds_identifier', index = 0)
            query_results_table.add_column(00000000                        , name='wds_rownum', index = 1)
            query_results_table.add_column('                              ', name = 'source_id_a', index = 2)
            query_results_table.add_column('                              ', name = 'source_id_b', index = 3)
```

astropy does not NOW allow adding columns as strings or a number. An attribute error arose. As the columns already existnig were added as a dictionary I suspected that a format description of some sort was needed. I used the add_column method and added a Table.Column() class that I created for each of the WDS related columns. These columns may be thought of column vectors with descriptions and these are accepted by the add_column() method.

This is an intermediate fix as an error is encountered later within ``` wds_in_gaia_query()``` as it loops through the WDS.