FOR Loop -> vector
Took Daphne's 06-16-22_build_and_write_astropy_table.py code and thought I could change the FOR loops to a vector method to speed things up. First try worked and this file includes a Jupyter notebook with the time comparisons that will print out. The FOR loop that calculates the secondary RA and DEC from the offset from the primary's RA and DEC is rewritten as a vector computation. The FOR loop took 390 seconds, the vector method less than a second.