Prototype version of gaia_wds.py
Version 0.1.x: creates the WDS table in VOTable, CSV and ECSV formats that are suitable for querying the ESA Gaia server. On 10 Jun 2024 I successfully queried Gaia DR3 and produced JOINs against the primaries and secondaries. Code takes 5 minutes or a little less to run. A big improvement over the 40-50 minutes that the FOR loop versions required to run. Gaia JOINs on the ESA server each took 21 seconds.